api/requirements.txt
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
click==8.1.3
Deprecated==1.2.13
exceptiongroup==1.1.1
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
limits==3.5.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
marshmallow-validators==3.0.0
mdurl==0.1.2
mistune==2.0.5
ordered-set==4.1.0
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.3.1
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.5
rich==13.3.5
six==1.16.0
SQLAlchemy==2.0.15
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.4
wrapt==1.15.0
